+++ src/components/common/whyrano/Tag.tsx
@@ -1,4 +1,4 @@
-import tagBox from '@components/common/tag/tag.css';
+import tagBox from '@components/common/whyrano/tag.css';
 
 interface TagProps {
   label: string;
